moved people test files to a new folder to group as a test suite unique location per resource tests set
