+++ src/test/java/tests/people/PeopleSmokeTests.java
@@ -1,13 +1,14 @@
-package tests;
+package tests.people;
 
 import java.io.IOException;
-import org.junit.jupiter.api.*;
 import helpers.config.MappingReader;
-import org.springframework.http.HttpMethod;
+import tests.BaseTests;
 import views.PeopleView;
+import org.junit.jupiter.api.*;
+import org.springframework.http.HttpMethod;
 
 @TestMethodOrder(MethodOrderer.OrderAnnotation.class)
-class PeopleSmokeTests extends BaseTests {
+public class PeopleSmokeTests extends BaseTests {
     private final PeopleView peopleView = new PeopleView();
     private final MappingReader mappingReader;
